scraper: Add company fetches and job adoption
Previously the steps were
	scrap -> nijobs -> accept
which lacked flexibility and was hard to reason about, so now
the steps are...
	fetch -> visit -> scrap -> link -> adopt -> nijobs -> accept
which correspond to make commands of the same name.

The previous architecture with only three steps _sucked balls_
as fetching and scraping had too many responsibilities and I could
not wrap by head around it. Too many jobs ended up with no parent
companies and there was no simple way to fix the problem without
fetching all the companies (there are over 9000 of them). Also,
creating stub companies was a _horrible_ idea. The stackoverflow
pages are not that hard to parse, so if we can't find some data
it's because it isn't there. So don't pretend, and discard any
pages we can't make sense of.
	- pages where jobs don't specify their parent company
	  stackoverflow uri path (/jobs/companies/*)
	- empty listings

So, break the arch down into seven steps - harder to grasp at first,
obviously overdesigned, but a lot more fixable and modifiable once the
data needed by nijobs expands or stackoverflow starts spewing
different pages (as it did just three days ago...)

- blobs.js: Write files to out/ folder etc.
- config.js: Read config from .env file.
- convert.js: Convert from stackoverflow format to nijobs format.
- fetch.js: Fetch stackoverflow pages and write them to out/ folder.
- files.js: Read and write data from/to files in different formats.
- html.js: Parse StackOverflow pages - css selector land.
- index.js: Run the make scripts.
- link.js: Link jobs and companies and write them to out/ folder.
- nijobs.js: Read linked data and write nijobs data to out/ folder.
- progress.js: Track progress and log information to terminal.
- random.js: Generate random data for missing, unimportant fields.
- roles.js: Handle developer roles - these are the dr parameters in
	StackOverflow and FieldType in nijobs.
- scrap.js: Read html data and write raw data to out/ folder.
- strings.js: String cleanup utility, regex land.
- tags.js: Handle StackOverflow tags and convert them to
	nijobs's TechnologyType.
- tidy.js: Validate and tidy scraps from html.js, and add cross
	references between jobs and companies.
- utils.js: Minimal lodash.
- visit.js: Read listing pages and write html pages to out/ folder.

- adopt.js: Visit companies needed for orphan jobs, scrap them,
	and link them with the orphan jobs. One function that
	does way too many things but is quite necessary for
	sane, connected and reliable data.